Don't rebuild CSVs multiple times in the same day
This will make it easier to play around with visualization options.

+++ lib/analytics/histogram_plotter.rb
@@ -33,6 +33,7 @@ class HistogramPlotter
 
   def build_csv
     FileUtils.mkdir_p analytics_path
+    return if File.exist? csv_path
     csv_content = CampaignArticlesCsvBuilder.new(@campaign).articles_to_csv
     File.write(csv_path, csv_content)
   end
